requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
google-generativeai==0.7.2
httpx==0.27.0
orjson==3.10.7
